Reindex anomalous HKL in the configured int dtype
```_reindex_per_reflection``` clones the new HKL table, which carries the
configured int dtype, and writes the stored anomalous HKL rows into it.
Under ``TORCHREF_DTYPE_INT=int64`` those rows were int32, and indexed
assignment rejects the mismatch; this predates the dtype conversion and
showed up in the int64 run of the reflection-data tests. Cast the rows to
the table's dtype.

+++ torchref/io/datasets/reflection_data.py
@@ -448,7 +448,7 @@ class ReflectionData(CrystalDataset, DebugMixin):
                 # Present rows keep their signed (anomalous) index; missing rows
                 # fall back to the canonical reference HKL (never a 0,0,0 row).
                 out = new_hkl.clone()
-                out[present] = val[src_idx]
+                out[present] = val[src_idx].to(out.dtype)
             else:
                 fill = self._REINDEX_FILL.get(name, 0)
                 out = torch.full(
